chore(refactor): methods/bankwirecard remove unused code
