--- packages/blockchain-wallet-v4-frontend/src/modals/BuySell/PaymentMethods/Methods/BankWireCard/BankWireCard.types.ts
@@ -1,11 +1,0 @@
-import { FC, ReactElement } from 'react'
-
-import { BSPaymentMethodType } from '@core/types'
-
-export type BankWireCardProps = {
-  onClick: (string) => void
-  text: ReactElement | string
-  value: BSPaymentMethodType
-}
-
-export type BankWireCardComponent = FC<BankWireCardProps>
